updated: model for blog and controller

+++ server/src/controllers/blog.controller.ts
@@ -2,6 +2,9 @@ import { Request, Response } from "express";
 import fs from "fs";
 import Blog, { IBlog } from "../models/blog";
 import User, { IUserModel } from "../models/user";
+import Tag, { ITag } from "../models/tag";
+import { Types } from "mongoose";
+import { titleToSlug } from "../utils/slug";
 
 export const getAllBlogs = async (
   req: Request,
@@ -10,7 +13,8 @@ export const getAllBlogs = async (
   try {
     const blogs: IBlog[] = await Blog.find()
       .populate("author", "_id name username")
-      .populate("comments");
+      .populate("comments")
+      .populate("tags");
 
     if (blogs.length < 1) {
       res.status(404).json({ message: "No posts yet." });
@@ -31,7 +35,8 @@ export const getBlogOfUser = async (
   try {
     const blogs: IBlog[] | null = await Blog.find({ author: userId })
       .populate("author", "_id name username")
-      .populate("comments");
+      .populate("comments")
+      .populate("tags");
 
     if (blogs.length < 1) {
       res.status(404).json({ message: "No posts yet." });
@@ -48,11 +53,17 @@ export const getBlogById = async (
   req: Request,
   res: Response
 ): Promise<void> => {
-  const { id } = req.params;
+  const { slug } = req.params;
+  console.log(slug);
+  const extractId = slug?.split("-");
+  const id = extractId[extractId?.length - 1];
+  console.log(id);
+
   try {
-    const blog: IBlog | null = await Blog.findById(id)
+    const blog: IBlog | null = await Blog.findOne({ _id: id })
       .populate("author", "_id name username")
-      .populate("comments");
+      .populate("comments")
+      .populate("tags");
 
     if (!blog) {
       res.status(404).json({ message: "Blog not found" });
@@ -69,13 +80,22 @@ export const createBlog = async (
   res: Response
 ): Promise<void> => {
   const { userId } = req.params;
-  let { cover, title, body, tags } = req.body;
+  const { title, body, tags } = req.body;
+  let { cover } = req.body;
 
   try {
-    if (!req.file) {
-      res.status(404).json({ message: "Cover image must have" });
+    const existingBlog: IBlog | null = await Blog.findOne({ title });
+
+    if (existingBlog) {
+      res.status(400).json({ message: "Title must be unique" });
       return;
     }
+
+    if (!req.file) {
+      res.status(400).json({ message: "Cover image must be provided" });
+      return;
+    }
+
     // Find the user by ID
     const user: IUserModel | null = await User.findById(userId);
     if (!user) {
@@ -83,24 +103,57 @@ export const createBlog = async (
       return;
     }
 
+    // Generate the _id for the new blog
+    const newBlogId = new Types.ObjectId().toHexString();
+
+    const slug = `${titleToSlug(title)}-${newBlogId}`;
+
+    // Update the cover image to use the filename from multer
     if (req.file) {
       cover = req.file.filename;
     }
 
+    // Find or create tag documents based on the tag names
+    const tagDocuments = await Promise.all(
+      tags.map(async (tagName: string) => {
+        const existingTag = await Tag.findOne({ name: tagName });
+
+        if (existingTag) {
+          return existingTag;
+        }
+
+        // If the tag doesn't exist, create a new tag
+        const newTag: ITag = new Tag({
+          name: tagName,
+        });
+        return newTag.save();
+      })
+    );
+
+    // Create a new blog using the Blog model
     const newBlog: IBlog = new Blog({
+      _id: newBlogId, // Assign the _id generated earlier
       author: userId,
       cover,
-      title,
+      title: title.trim(),
       body,
-      tags,
+      tags: tagDocuments.map((tag) => tag._id), // Store only the tag IDs in the blog's tags array
+      slug,
     });
 
     const savedBlog: IBlog = await newBlog.save();
 
+    // Update the tag documents to include the newly created blog in their blogs array
+    tagDocuments.forEach(async (tag) => {
+      tag.blogs.push(savedBlog._id);
+      await tag.save();
+    });
+
     res
       .status(200)
-      .json({ message: "created successfully", result: savedBlog });
+      .json({ message: "Created successfully", result: savedBlog });
   } catch (error) {
+    console.error("Error creating blog:", error);
     res.status(500).json({ message: "Server Error" });
   }
 };
@@ -113,25 +166,58 @@ export const updateBlog = async (
   let { title, body, tags, cover, fileOld } = req.body;
 
   try {
+    // const slug = `${title.trim().toLowerCase().replaceAll(" ", "-")}-${id}`;
+    const slug = `${titleToSlug(title)}-${id}`;
+
+    // Check if the new title is unique (excluding the current blog being updated)
+    const existingBlog: IBlog | null = await Blog.findOne({
+      title: { $regex: new RegExp("^" + title + "$", "i") },
+      _id: { $ne: id },
+    });
+
+    if (existingBlog) {
+      res.status(400).json({ message: "Title must be unique" });
+      return;
+    }
+
+    // If there is a new uploaded file, update the cover and delete the old image
     if (req.file) {
       cover = req.file.filename;
       const imagePath = `public/images/${fileOld}`;
       fs.unlinkSync(imagePath);
     }
 
+    // Convert the provided tag names into tag IDs
+    const tagIds: string[] = [];
+    for (const tagName of tags) {
+      const tag: ITag | null = await Tag.findOne({ name: tagName });
+      if (tag) {
+        tagIds.push(tag._id);
+      } else {
+        // If the tag doesn't exist, create a new one and get its ID
+        const newTag: ITag = new Tag({ name: tagName });
+        const savedTag: ITag = await newTag.save();
+        tagIds.push(savedTag._id);
+      }
+    }
+
+    // Find the blog by ID and update its fields
     const updatedBlog: IBlog | null = await Blog.findByIdAndUpdate(
       id,
-      { cover, title, body, tags },
+      { cover, title: title.trim(), body, tags: tagIds, slug },
       { new: true }
     );
+
     if (!updatedBlog) {
       res.status(404).json({ message: "Blog not found" });
+      return;
     }
 
     res
       .status(200)
-      .json({ message: "updated successfully", result: updatedBlog });
+      .json({ message: "Updated successfully", result: updatedBlog });
   } catch (error) {
+    console.error("Error updating blog:", error);
     res.status(500).json({ message: "Server Error" });
   }
 };
@@ -154,6 +240,15 @@ export const deleteBlog = async (
       const imagePath = `public/images/${deletedBlog.cover}`;
       fs.unlinkSync(imagePath);
     }
+
+    // Remove the deleted blog's ID from the 'blogs' array of associated tags
+    const tagsToUpdate: ITag[] = await Tag.find({ blogs: id });
+    const updateOperations = tagsToUpdate.map((tag) => {
+      tag.blogs = tag.blogs.filter((blogId) => blogId.toString() !== id);
+      return tag.save();
+    });
+
+    await Promise.all(updateOperations);
 
     res.status(200).json({ message: "deleted successfully" });
   } catch (error) {
